+++ src/math/activations/sigmoid.rs
@@ -1,8 +1,8 @@
 
-pub fn sigmoid(x: f64) -> f64 {
+pub fn sigmoid(x: f32) -> f32 {
     1.0 / (1.0 + (-x).exp())
 }
 
-pub fn sigmoid_derivative(x: f64) -> f64 {
+pub fn sigmoid_derivative(x: f32) -> f32 {
     x * (1.0 - x)
 }
