+++ src/math/activations/softmax.rs
@@ -1,11 +1,11 @@
-use std::f64::consts::E;
+use std::f32::consts::E;
 use crate::Vector;
 
-pub fn softmax(vec: &Vec<f64>) -> Vec<f64> {
-    let mut result: Vec<f64> = Vec::new();
+pub fn softmax(vec: &Vec<f32>) -> Vec<f32> {
+    let mut result: Vec<f32> = Vec::new();
     let mut sum = 0.0;
     for i in 0..vec.len() {
-        let exp_val: f64 = E.powf(vec[i]);
+        let exp_val: f32 = E.powf(vec[i]);
         result.push(exp_val);
         sum += exp_val;
     }
@@ -15,8 +15,8 @@ pub fn softmax(vec: &Vec<f64>) -> Vec<f64> {
     result
 }
 
-pub fn softmax_derivative(vec: &Vec<f64>) -> Vec<f64> {
-    let mut result: Vec<f64> = Vec::new();
+pub fn softmax_derivative(vec: &Vec<f32>) -> Vec<f32> {
+    let mut result: Vec<f32> = Vec::new();
     for i in 0..vec.len() {
         let mut sum = 0.0;
         for j in 0..vec.len() {
