+++ src/math/activations/tanh.rs
@@ -1,8 +1,8 @@
 
-pub fn tanh(x: f64) -> f64 {
+pub fn tanh(x: f32) -> f32 {
     x.tanh()
 }
 
-pub fn tanh_derivative(x: f64) -> f64 {
+pub fn tanh_derivative(x: f32) -> f32 {
     1.0 - x.powi(2)
 }
